README.md
This project aims to create a data pipeline for analyzing stock price data. The pipeline fetches data from financial APIs, processes it using Apache Beam, stores it in a MySQL database, and performs basic analysis and visualization using Python in Google Colab. The focus of the analysis is on volume data to understand the impact of stock market reports.